feat: serialisation for transfers for out of band sending
