+++ sn_transfers/src/wallet/utxo.rs
@@ -6,9 +6,11 @@
 // KIND, either express or implied. Please review the Licences for the specific language governing
 // permissions and limitations relating to use of the SAFE Network Software.
 
+use std::collections::BTreeMap;
+
 use bls::{Ciphertext, PublicKey, SecretKey};
 use serde::{Deserialize, Serialize};
-use sn_dbc::DerivationIndex;
+use sn_dbc::{Dbc, DerivationIndex, PublicAddress};
 use sn_protocol::storage::DbcAddress;
 
 use super::error::{Error, Result};
@@ -22,6 +24,38 @@ pub struct Transfer {
 }
 
 impl Transfer {
+    /// Creates Transfers from the given dbcs
+    /// Grouping DBCs by recipient into different transfers
+    /// This Transfer can be sent safely to the recipients as all data in it is encrypted
+    /// The recipients can then decrypt the data and use it to verify and reconstruct the DBCs
+    pub fn transfers_from_dbcs(dbcs: Vec<Dbc>) -> Result<Vec<Transfer>> {
+        let mut utxos_map: BTreeMap<PublicAddress, Vec<Utxo>> = BTreeMap::new();
+        for dbc in dbcs {
+            let recipient = dbc.secrets.public_address;
+            let derivation_index = dbc.derivation_index();
+            let parent_spend_addr = match dbc.signed_spends.iter().next() {
+                Some(s) => DbcAddress::from_dbc_id(s.dbc_id()),
+                None => {
+                    warn!(
+                        "Skipping DBC {dbc:?} while creating Transfer as it has no parent spends."
+                    );
+                    continue;
+                }
+            };
+
+            let u = Utxo::new(derivation_index, parent_spend_addr);
+            utxos_map.entry(recipient).or_insert_with(Vec::new).push(u);
+        }
+
+        let mut transfers = Vec::new();
+        for (recipient, utxos) in utxos_map {
+            let t =
+                Transfer::create(utxos, recipient.0).map_err(|_| Error::UtxoEncryptionFailed)?;
+            transfers.push(t)
+        }
+        Ok(transfers)
+    }
+
     /// Create a new transfer
     /// utxos: List of UTXOs to be used for payment
     /// recipient: main Public key (donation key) of the recipient,
@@ -43,6 +77,23 @@ impl Transfer {
             utxos.push(utxo);
         }
         Ok(utxos)
+    }
+
+    /// Deserializes a `Transfer` represented as a hex string to a `Transfer`.
+    pub fn from_hex(hex: &str) -> Result<Self> {
+        let mut bytes = hex::decode(hex).map_err(|_| Error::TransferDeserializationFailed)?;
+        bytes.reverse();
+        let transfer: Transfer =
+            bincode::deserialize(&bytes).map_err(|_| Error::TransferDeserializationFailed)?;
+        Ok(transfer)
+    }
+
+    /// Serialize this `Transfer` instance to a hex string.
+    pub fn to_hex(&self) -> Result<String> {
+        let mut serialized =
+            bincode::serialize(&self).map_err(|_| Error::TransferSerializationFailed)?;
+        serialized.reverse();
+        Ok(hex::encode(serialized))
     }
 }
